changed some details for wh65b

+++ LaCrosseITPlusReader10/WSBase.h
@@ -34,7 +34,7 @@ public:
     float  WindSpeed;     // m/s
     float  WindGust;      // m/s
     int  Pressure;        // hPa
-    float UV;
+    int UV;
     float Light;          // lux
     int16_t strikesDistance ;
     uint16_t strikesTotal ;
